pyproject.toml
[project]
name = "anki-cards-from-ai-chats"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.56.0",
    "requests>=2.32.5",
]
